Move test for mock API to test_mgx_api.py
