+++ tests/me/test_metagenomics_exchange.py
@@ -2,14 +2,10 @@
 # -*- coding: utf-8 -*-
 
 import pytest
-from unittest import mock
 
 from emgapi.metagenomics_exchange import MetagenomicsExchangeAPI
 
 import requests
-import responses
-import settings
-
 
 class TestME:
     def test_check_existing_analysis(self):
@@ -28,19 +24,3 @@ class TestME:
         # Should return -> https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/409
         with pytest.raises(requests.HTTPError, match="409 Client Error"):
             me_api.add_record(mgya=source_id, run_accession="ERR3063408", public=True).json()
-
-    @responses.activate
-    def test_mock_post_new_analysis(self):
-        me_api = MetagenomicsExchangeAPI()
-        endpoint = "datasets"
-        url = settings.ME_API + f"/{endpoint}"
-
-        responses.add(responses.POST, url, json={'success': True}, status=201)
-
-        response = me_api.add_record(mgya="MGYA00593709", run_accession="SRR3960575", public=True)
-
-        assert response.status_code == 201
-        assert response.json() == {'success': True}
-
-
-
