+++ tests/me/test_mgx_api.py
@@ -15,13 +15,30 @@
 # limitations under the License.
 
 import pytest
+import responses
 from unittest.mock import patch
 
 from django.core.management import call_command
+import settings
 
 from test_utils.emg_fixtures import *  # noqa
 
+from emgapi.metagenomics_exchange import MetagenomicsExchangeAPI
 from emgapi.models import AnalysisJob, MetagenomicsExchange, ME_Broker
+
+class TestMockResponse:
+    @responses.activate
+    def test_mock_post_new_analysis(self):
+        me_api = MetagenomicsExchangeAPI()
+        endpoint = "datasets"
+        url = settings.ME_API + f"/{endpoint}"
+
+        responses.add(responses.POST, url, json={'success': True}, status=201)
+
+        response = me_api.add_record(mgya="MGYA00593709", run_accession="SRR3960575", public=True)
+
+        assert response.status_code == 201
+        assert response.json() == {'success': True}
 
 @pytest.mark.django_db
 class TestMeAPI:
